Beginning of the Save Game Manager
Contains the first screen where the user can create a new game or load a save game. Creating a new game works and viewing the newly created games in the list works (on game restart). Some of the code works; mostly the SaveGameManager's Get and Update functions are not tested. The Save button will eventually do some error checking and take the user to the next screen. The Load button will eventually do more than nothing.